Added more classes still in progress
Character, Formation, and Skill are involved with outside the Battle
Simulation - however, this creates a link between the Battle class and
with the Character which will essentially be connected to the World.

Unit is the instantiation of the character (with current health, ect.)
Group is the instantiation of the formation (simulated in battles and
exists in world)

Battle is the simulator between two Groups.