Modifications to Battle Simulator
Downgraded Visual Studio 2010

Augmented Ability class with Action Types. Standard Actions are executed
during a unit's turn (the first skill). There will be Response Action
Abilities that apply to things such as counterattacks. Secondly, added a
Status (Buff) effect called StatusAttackResponse, which is a buff that
can be linked to an ability that will be executed in response to an
attack. In addition, variables are added in preparation for this to both
Status and Ability class such as interruptible and preemptive.

Note to Self: Status Buffs/Debuffs must be checked if they are expired
on call so that expired Buffs/Debuffs are not reapplied again due to
efficient cleaning.