Base Framework is Prepared
Major additions and changes:
- Project and output is cleaned.
- Debugged and fixed Status effects and their stacking/reset behavior
(Mortality, Bshout, ect.)
- Heal skill is implemented and tested (Heals 10 when used by a fighter)
- Counterattacks are implemented (actually called Response abilities,
because you can respond to an attack with a Heal spell). Determines
whether an ability is respondable and if a primary target is set and
provoked. If so, a response ability is generated and executes. It is
currently set so that melee attacks counter melee attacks and ranged
attacks counter ranged attacks. A unit has three skill tiers to choose a
skill from. The first tier is NO_STANDARD_SKILL, the second tier is a
BASIC_SKILL, and the third tier is the ROW_SKILL. If a unit is stunned,
ect., its tier is set to the first tier. In general, a standard skill
will be set initially to the third tier unless a status effect forces it
down to a lower level. This makes it easy so that a response ability can
be set to the second tier while giving the ability to still be
downgraded to NO_SKILL tier.

Todo:
- Resolve some Counter-attack questions in Project Ogre chat and tweak
it based on result of discussion
- Implement and test more Classes
- Implement and test more Skills (Next up, add a unbuff or dispel skill
and see if it breaks)
- More distinguishing ability types (currently at ATTACK_MELEE,
ATTACK_RANGE, SPECIAL)
- Use some graphical output
- Have a nice distinguish on different types of buff stacks can do. Such
as a stackable buff, but each buff has its own timer still. Or the
stackable buff that resets the timer. Or a resetting buff.
- write functionality to make HOLY or SHADOW damage easy in programming
(they are composed of three elements)
- This may not be the case for this game, but if it is, add a
distinguish between evade, resist, ect. when determining chance of
success for a hit. Furthermore, we can add accuracy and magic
penetration later.